validate uniqueness of uids
Clusters have been broken several time because we added users with uids
that overlapped with existing uids.

Here we add a simple validation that causes ./install.sh to abort if
non-unique uids are detected in the config file.

Note: ./install.sh is also called by ./add_users.sh, i.e. it covers both
the case of new deployments and that of adding users to existing
deployments.